refactor(edit-engine): split transaction domains
Extract connectivity evidence, Net lifecycle, physical-contact, and MOS bulk reconciliation helpers; move formal Cell interface edits behind one typed dispatcher; and isolate presentation, annotation, drafting, and layout edits behind a second typed dispatcher.

executeTransaction remains the ordering and rejection coordinator while transaction.ts falls from 4,246 to 2,993 lines. Each extracted domain receives only its draft, rejection seam, and change-tracking capabilities.

Validation: pnpm ci:static; complete edit-engine unit suite (236); complete hierarchy E2E (14) after connectivity extraction; focused drafting (3) and Cell-interface hierarchy (3) E2E after domain dispatch extraction; git diff --check.

No test files changed because existing transaction, drafting, presentation, and hierarchy contracts directly exercise the preserved edit semantics.

Unrelated user-owned probe-conflicts.mjs remained untracked and untouched.

Test-Impact: no-test-change — behavior-preserving transaction domain extraction covered by 236 package tests and 20 browser scenarios.
