+++ packages/edit-engine/src/transaction.ts
@@ -1,26 +1,14 @@
 import {
   AnnotationSchema,
-  CellNetlistTerminalSchema,
   ConnectivityEvidenceSchema,
-  DraftingObjectSchema,
   InstanceSchema,
   JunctionSchema,
-  LayoutConstraintSchema,
-  LayoutGroupSchema,
   NoConnectSchema,
   SchematicDocumentSchema,
   deriveStableId,
-  foldNetName,
 } from "@icm/model";
 import { createReferenceIndex, referenceIssuesForInstance } from "@icm/devices";
-import type {
-  Annotation,
-  Point,
-  Rotation,
-  RouteBranch,
-  RouteEndpoint,
-  SchematicDocument,
-} from "@icm/model";
+import type { RouteBranch, SchematicDocument } from "@icm/model";
 import {
   endpointKey,
   hasExplicitMosBulkRoute,
@@ -33,7 +21,6 @@ import {
   resolveMosBulkConnection,
   validateLogicalNetContract,
 } from "@icm/derived";
-import type { SymbolResolver } from "@icm/symbols";
 import { EditTransactionSchema, type EditTransaction } from "./edit-schema.js";
 import {
   buildOrthogonalEscapeRoute,
@@ -60,22 +47,38 @@ import {
 import {
   followAttachedAnnotations,
   refreshInstanceValueAnnotation,
-  translateObjectAnchoredAnnotation,
 } from "./transaction-instance-annotations.js";
 import { reconcileTransformDirectContacts } from "./transaction-direct-contact.js";
 import { nextPhysicalContactOperation } from "./transaction-connectivity-normalizer.js";
+import { applyCellInterfaceEdit } from "./transaction-cell-interface.js";
+import { applyPresentationLayoutEdit } from "./transaction-presentation-layout.js";
+import {
+  type BulkDefaultIdentity,
+  connectivityEvidenceNetIds,
+  implicitBulkPresentation,
+  mergeBaseNets,
+  physicalContactObjectIdsForTransaction,
+  preferredPhysicalMergeTarget,
+  propagateSpiceSourceEvidenceAfterSplit,
+  pruneUnreachableLocalNet,
+  reconcileMaterializedMosBulkBindings,
+  removeConnectivityEvidenceOwnedBy,
+  removeNoConnectForEndpoint,
+  retargetConnectivityEvidenceOwner,
+  retargetMosBulkDefaultsAfterSplit,
+  retargetOwnerEvidenceAfterSplit,
+  revokeInvalidatedSupplyBulkDefaults,
+  uniquePhysicalContactId,
+} from "./transaction-connectivity.js";
 import {
   addEndpointToNet,
   endpointOwnerNetId,
   lockedLayoutOwner,
   netEndpointGroups,
-  pointOnSegment,
-  replaceLayoutReference,
   routeFromEdit,
   routeIsProtected,
   sameResolvedRoutePoints,
   validateConnectableEndpoint,
-  validateNetLabelBinding,
   validateRoute,
 } from "./transaction-routing.js";
 import {
@@ -116,865 +119,6 @@ function referencePolicyFailure(
     case "DUPLICATE_REFERENCE":
       return `Reference ${issue.reference} is already used by ${issue.otherInstanceId}`;
   }
-}
-
-function retargetConnectivityEvidence(
-  draft: SchematicDocument,
-  sourceNetId: string,
-  targetNetId: string,
-  changedObjectIds: Set<string>,
-): void {
-  const retainedEvidence: typeof draft.connectivityEvidence = [];
-  for (const evidence of draft.connectivityEvidence) {
-    if (evidence.kind === "explicit-equivalence") {
-      const memberNetIds = [
-        ...new Set(
-          evidence.memberNetIds.map((netId) =>
-            netId === sourceNetId ? targetNetId : netId,
-          ),
-        ),
-      ];
-      if (memberNetIds.length < 2) {
-        changedObjectIds.add(evidence.id);
-        continue;
-      }
-      if (
-        memberNetIds.length !== evidence.memberNetIds.length ||
-        memberNetIds.some(
-          (netId, index) => netId !== evidence.memberNetIds[index],
-        )
-      ) {
-        evidence.memberNetIds = memberNetIds;
-        changedObjectIds.add(evidence.id);
-      }
-    } else if (evidence.netId === sourceNetId) {
-      evidence.netId = targetNetId;
-      changedObjectIds.add(evidence.id);
-    }
-    retainedEvidence.push(evidence);
-  }
-  const seenSpiceSources = new Set<string>();
-  draft.connectivityEvidence = retainedEvidence.filter((evidence) => {
-    if (evidence.kind !== "spice-source") return true;
-    const key = `${evidence.netId}\u0000${evidence.sourceNetId}`;
-    if (!seenSpiceSources.has(key)) {
-      seenSpiceSources.add(key);
-      return true;
-    }
-    changedObjectIds.add(evidence.id);
-    return false;
-  });
-}
-
-type BaseNetMergeResult =
-  | { ok: true }
-  | {
-      ok: false;
-      code: "OBJECT_NOT_FOUND" | "EDIT_PRECONDITION";
-      message: string;
-      netIds: readonly string[];
-    };
-
-function mergeBaseNets(
-  draft: SchematicDocument,
-  targetNetId: string,
-  sourceNetId: string,
-  changedObjectIds: Set<string>,
-): BaseNetMergeResult {
-  if (targetNetId === sourceNetId) return { ok: true };
-  const target = draft.nets.find((net) => net.id === targetNetId);
-  const sourceIndex = draft.nets.findIndex((net) => net.id === sourceNetId);
-  const source = draft.nets[sourceIndex];
-  if (!target || !source) {
-    return {
-      ok: false,
-      code: "OBJECT_NOT_FOUND",
-      message: `Net merge target/source does not exist: ${targetNetId}, ${sourceNetId}`,
-      netIds: [targetNetId, sourceNetId],
-    };
-  }
-  const logicalNets = resolveDocumentLogicalNets(draft);
-  const targetPowerDomain =
-    logicalNets.byBaseNetId.get(target.id)?.powerDomain ?? "none";
-  const sourcePowerDomain =
-    logicalNets.byBaseNetId.get(source.id)?.powerDomain ?? "none";
-  if (
-    targetPowerDomain === "conflict" ||
-    sourcePowerDomain === "conflict" ||
-    (targetPowerDomain !== "none" &&
-      sourcePowerDomain !== "none" &&
-      targetPowerDomain !== sourcePowerDomain)
-  ) {
-    return {
-      ok: false,
-      code: "EDIT_PRECONDITION",
-      message: "Cannot merge Nets with incompatible power domains",
-      netIds: [target.id, source.id],
-    };
-  }
-  for (const instance of draft.instances) {
-    if (instance.mosBulkBinding?.netId === source.id) {
-      instance.mosBulkBinding.netId = target.id;
-      changedObjectIds.add(instance.id);
-    }
-  }
-  if (draft.mosBulkDefaults?.nmosNetId === source.id) {
-    draft.mosBulkDefaults.nmosNetId = target.id;
-  }
-  if (draft.mosBulkDefaults?.pmosNetId === source.id) {
-    draft.mosBulkDefaults.pmosNetId = target.id;
-  }
-  for (const terminal of draft.netlist?.terminals ?? []) {
-    if (terminal.netId === source.id) terminal.netId = target.id;
-  }
-  for (const terminal of source.terminals) {
-    if (
-      !target.terminals.some(
-        (candidate) =>
-          candidate.instanceId === terminal.instanceId &&
-          candidate.pinName === terminal.pinName,
-      )
-    ) {
-      target.terminals.push(structuredClone(terminal));
-    }
-  }
-  for (const route of draft.routes) {
-    if (route.netId === source.id) {
-      route.netId = target.id;
-      changedObjectIds.add(route.id);
-    }
-  }
-  for (const junction of draft.junctions) {
-    if (junction.netId === source.id) {
-      junction.netId = target.id;
-      changedObjectIds.add(junction.id);
-    }
-  }
-  for (const annotation of draft.annotations) {
-    if (annotation.netId === source.id) {
-      annotation.netId = target.id;
-      changedObjectIds.add(annotation.id);
-    }
-    if (
-      annotation.binding?.kind === "net-name" &&
-      annotation.binding.netId === source.id
-    ) {
-      annotation.binding = { kind: "net-name", netId: target.id };
-      changedObjectIds.add(annotation.id);
-    }
-  }
-  for (const group of draft.layoutGroups) {
-    const replaced = group.objectIds.includes(source.id);
-    group.objectIds = replaceLayoutReference(
-      group.objectIds,
-      source.id,
-      target.id,
-    );
-    if (replaced) changedObjectIds.add(group.id);
-  }
-  for (const constraint of draft.constraints) {
-    const replaced = constraint.objectIds.includes(source.id);
-    constraint.objectIds = replaceLayoutReference(
-      constraint.objectIds,
-      source.id,
-      target.id,
-    );
-    if (replaced) changedObjectIds.add(constraint.id);
-  }
-  retargetConnectivityEvidence(draft, source.id, target.id, changedObjectIds);
-  draft.nets.splice(sourceIndex, 1);
-  changedObjectIds.add(target.id);
-  changedObjectIds.add(source.id);
-  return { ok: true };
-}
-
-function removeNoConnectForEndpoint(
-  draft: SchematicDocument,
-  endpoint: RouteEndpoint,
-  changedObjectIds: Set<string>,
-): void {
-  if (endpoint.kind !== "terminal") return;
-  draft.noConnects = draft.noConnects.filter((noConnect) => {
-    const matches =
-      noConnect.endpoint.instanceId === endpoint.instanceId &&
-      noConnect.endpoint.pinName === endpoint.pinName;
-    if (matches) changedObjectIds.add(noConnect.id);
-    return !matches;
-  });
-}
-
-function preferredPhysicalMergeTarget(
-  draft: SchematicDocument,
-  leftNetId: string,
-  rightNetId: string,
-): readonly [targetNetId: string, sourceNetId: string] {
-  const logical = resolveDocumentLogicalNets(draft);
-  const score = (netId: string): number => {
-    const resolved = logical.byBaseNetId.get(netId);
-    return (
-      (resolved?.powerDomain !== undefined &&
-      resolved.powerDomain !== "none" &&
-      resolved.powerDomain !== "conflict"
-        ? 4
-        : 0) +
-      (resolved?.scope === "global" ? 2 : 0) +
-      (resolved?.name ? 1 : 0)
-    );
-  };
-  const leftScore = score(leftNetId);
-  const rightScore = score(rightNetId);
-  if (leftScore !== rightScore) {
-    return leftScore > rightScore
-      ? [leftNetId, rightNetId]
-      : [rightNetId, leftNetId];
-  }
-  return leftNetId.localeCompare(rightNetId, "en") <= 0
-    ? [leftNetId, rightNetId]
-    : [rightNetId, leftNetId];
-}
-
-function uniquePhysicalContactId(
-  draft: SchematicDocument,
-  kind: "net" | "route",
-  transactionId: string,
-  seed: string,
-): string {
-  const occupied = new Set([
-    ...draft.instances.map((object) => object.id),
-    ...draft.nets.map((object) => object.id),
-    ...draft.routes.map((object) => object.id),
-    ...draft.junctions.map((object) => object.id),
-    ...draft.noConnects.map((object) => object.id),
-    ...draft.annotations.map((object) => object.id),
-    ...draft.connectivityEvidence.map((object) => object.id),
-    ...draft.layoutGroups.map((object) => object.id),
-    ...draft.constraints.map((object) => object.id),
-    ...(draft.drafting?.objects.map((object) => object.id) ?? []),
-    ...(draft.netlist?.terminals.map((object) => object.id) ?? []),
-  ]);
-  let attempt = 0;
-  while (true) {
-    const id = deriveStableId(
-      kind,
-      draft.id,
-      "physical-contact",
-      transactionId,
-      seed,
-      String(attempt),
-    );
-    if (!occupied.has(id)) return id;
-    attempt += 1;
-  }
-}
-
-function physicalContactObjectIdsForTransaction(
-  transaction: EditTransaction,
-): Set<string> {
-  const result = new Set<string>();
-  for (const edit of transaction.edits) {
-    switch (edit.kind) {
-      case "add_instance":
-        result.add(edit.instance.id);
-        break;
-      case "set_instance_symbol":
-      case "place_instance":
-      case "move_instance":
-      case "rotate_instance":
-      case "mirror_instance":
-        result.add(edit.instanceId);
-        break;
-      case "align_instances":
-        edit.instanceIds.forEach((instanceId) => result.add(instanceId));
-        break;
-      case "set_route_points":
-      case "route_orthogonal":
-        result.add(edit.routeId);
-        break;
-      case "add_junction":
-      case "move_junction":
-        result.add(edit.junctionId);
-        break;
-      case "attach_endpoint_to_route":
-        result.add(
-          edit.endpoint.kind === "terminal"
-            ? edit.endpoint.instanceId
-            : edit.endpoint.junctionId,
-        );
-        result.add(edit.routeId);
-        break;
-      case "add_power_rail":
-        result.add(edit.routeId);
-        result.add(edit.startJunctionId);
-        result.add(edit.endJunctionId);
-        break;
-    }
-  }
-  return result;
-}
-
-function connectivityEvidenceNetIds(
-  evidence: SchematicDocument["connectivityEvidence"][number],
-): readonly string[] {
-  return evidence.kind === "explicit-equivalence"
-    ? evidence.memberNetIds
-    : [evidence.netId];
-}
-
-function connectivityEvidenceOwnerId(
-  evidence: SchematicDocument["connectivityEvidence"][number],
-): string | null {
-  if (evidence.kind !== "name-claim") return null;
-  switch (evidence.owner.kind) {
-    case "net-label":
-      return evidence.owner.annotationId;
-    case "power-marker":
-      return evidence.owner.objectId;
-    case "explicit-net-property":
-      return null;
-  }
-}
-
-function removeConnectivityEvidenceOwnedBy(
-  draft: SchematicDocument,
-  objectIds: ReadonlySet<string>,
-  changedObjectIds: Set<string>,
-): readonly string[] {
-  const affectedNetIds = new Set<string>();
-  const sourceBackedNetIds = new Set(
-    draft.connectivityEvidence.flatMap((evidence) =>
-      evidence.kind === "spice-source" ? [evidence.netId] : [],
-    ),
-  );
-  const removedOwnedClaims = draft.connectivityEvidence.filter((evidence) => {
-    const ownerId = connectivityEvidenceOwnerId(evidence);
-    return Boolean(ownerId && objectIds.has(ownerId));
-  });
-  const shadowedProjectionIds = new Set(
-    draft.connectivityEvidence.flatMap((evidence) => {
-      if (
-        evidence.kind !== "name-claim" ||
-        evidence.owner.kind !== "explicit-net-property" ||
-        sourceBackedNetIds.has(evidence.netId)
-      ) {
-        return [];
-      }
-      const shadowed = removedOwnedClaims.some(
-        (candidate) =>
-          candidate.kind === "name-claim" &&
-          candidate.netId === evidence.netId &&
-          foldNetName(candidate.name) === foldNetName(evidence.name) &&
-          candidate.scope === evidence.scope &&
-          candidate.powerDomain === evidence.powerDomain,
-      );
-      return shadowed ? [evidence.id] : [];
-    }),
-  );
-  draft.connectivityEvidence = draft.connectivityEvidence.filter((evidence) => {
-    const ownerId = connectivityEvidenceOwnerId(evidence);
-    if (
-      (!ownerId || !objectIds.has(ownerId)) &&
-      !shadowedProjectionIds.has(evidence.id)
-    ) {
-      return true;
-    }
-    changedObjectIds.add(evidence.id);
-    for (const netId of connectivityEvidenceNetIds(evidence)) {
-      affectedNetIds.add(netId);
-    }
-    return false;
-  });
-  return [...affectedNetIds];
-}
-
-function retargetConnectivityEvidenceOwner(
-  draft: SchematicDocument,
-  sourceObjectId: string,
-  targetObjectId: string,
-  changedObjectIds: Set<string>,
-): void {
-  for (const evidence of draft.connectivityEvidence) {
-    if (
-      evidence.kind === "name-claim" &&
-      evidence.owner.kind === "power-marker" &&
-      evidence.owner.objectId === sourceObjectId
-    ) {
-      evidence.owner.objectId = targetObjectId;
-      changedObjectIds.add(evidence.id);
-    }
-  }
-}
-
-function retargetOwnerEvidenceAfterSplit(
-  draft: SchematicDocument,
-  originalNetId: string,
-  netIdByEndpoint: ReadonlyMap<string, string>,
-  changedObjectIds: Set<string>,
-): void {
-  const instanceNetId = (instanceId: string): string | undefined =>
-    draft.nets
-      .flatMap((net) =>
-        net.terminals.some((terminal) => terminal.instanceId === instanceId)
-          ? [net.id]
-          : [],
-      )
-      .sort((left, right) => left.localeCompare(right, "en"))[0];
-  const objectNetId = (objectId: string): string | undefined =>
-    draft.routes.find((route) => route.id === objectId)?.netId ??
-    draft.junctions.find((junction) => junction.id === objectId)?.netId ??
-    draft.annotations.find((annotation) => annotation.id === objectId)?.netId ??
-    instanceNetId(objectId);
-
-  for (const evidence of draft.connectivityEvidence) {
-    if (evidence.kind !== "name-claim" || evidence.netId !== originalNetId) {
-      continue;
-    }
-    let targetNetId: string | undefined;
-    if (evidence.owner.kind === "net-label") {
-      const annotationId = evidence.owner.annotationId;
-      const annotation = draft.annotations.find(
-        (candidate) => candidate.id === annotationId,
-      );
-      if (annotation?.anchor.kind === "route") {
-        const routeId = annotation.anchor.routeId;
-        targetNetId = draft.routes.find((route) => route.id === routeId)?.netId;
-      } else if (annotation?.anchor.kind === "object") {
-        targetNetId = objectNetId(annotation.anchor.objectId);
-      }
-      if (annotation && targetNetId && targetNetId !== annotation.netId) {
-        annotation.netId = targetNetId;
-        if (annotation.binding?.kind === "net-name") {
-          annotation.binding = { kind: "net-name", netId: targetNetId };
-        }
-        changedObjectIds.add(annotation.id);
-      }
-    } else if (evidence.owner.kind === "power-marker") {
-      targetNetId = objectNetId(evidence.owner.objectId);
-    }
-    if (targetNetId && targetNetId !== evidence.netId) {
-      evidence.netId = targetNetId;
-      changedObjectIds.add(evidence.id);
-    }
-  }
-}
-
-function propagateSpiceSourceEvidenceAfterSplit(
-  draft: SchematicDocument,
-  originalNetId: string,
-  splitNetIds: readonly string[],
-  changedObjectIds: Set<string>,
-): void {
-  const sourceNetIds = draft.connectivityEvidence.flatMap((evidence) =>
-    evidence.kind === "spice-source" && evidence.netId === originalNetId
-      ? [evidence.sourceNetId]
-      : [],
-  );
-  for (const sourceNetId of new Set(sourceNetIds)) {
-    for (const netId of splitNetIds) {
-      if (
-        draft.connectivityEvidence.some(
-          (evidence) =>
-            evidence.kind === "spice-source" &&
-            evidence.netId === netId &&
-            evidence.sourceNetId === sourceNetId,
-        )
-      ) {
-        continue;
-      }
-      const id = deriveStableId(
-        "connectivity-evidence",
-        "spice-source",
-        sourceNetId,
-        netId,
-      );
-      draft.connectivityEvidence.push({
-        id,
-        kind: "spice-source",
-        netId,
-        sourceNetId,
-      });
-      changedObjectIds.add(id);
-    }
-  }
-}
-
-/**
- * Ensure the ADR 0010 drafting layer exists on a draft Document. It is
- * optional in the schema so legacy Projects still validate; edits that touch
- * drafting initialize an empty container first.
- */
-function ensureDraftingLayer(draft: SchematicDocument): void {
-  if (!draft.drafting) {
-    draft.drafting = { objects: [] };
-  }
-}
-
-/**
- * A local Net with no electrical or authored presentation reachability is
- * implementation debris, not a reusable electrical object. This is called
- * immediately after the final endpoint is disconnected so a later
- * `remove_instance` cannot retain a stale Port designator through its Net.
- *
- * Evidence describes a live Base Net; it cannot make an otherwise unreachable
- * Base Net live by referring back to it. Visible labels/markers are already
- * counted through their Annotation, Instance, Route, or Junction owners.
- * Imported source/name evidence and explicit equivalence are retired or
- * trimmed with the Base Net, while the Document source binding remains the
- * durable provenance record.
- */
-function pruneUnreachableLocalNet(
-  draft: SchematicDocument,
-  netId: string,
-  changedObjectIds: Set<string>,
-  options: {
-    deferInto?: Set<string>;
-    protectedEvidenceIds?: ReadonlySet<string>;
-  } = {},
-): void {
-  options.deferInto?.add(netId);
-  const net = draft.nets.find((candidate) => candidate.id === netId);
-  if (!net || net.terminals.length > 0) {
-    return;
-  }
-  if (
-    draft.routes.some((route) => route.netId === netId) ||
-    draft.junctions.some((junction) => junction.netId === netId) ||
-    draft.netlist?.terminals.some((terminal) => terminal.netId === netId) ||
-    draft.annotations.some(
-      (annotation) =>
-        annotation.netId === netId ||
-        (annotation.binding?.kind === "net-name" &&
-          annotation.binding.netId === netId),
-    ) ||
-    draft.layoutGroups.some((group) => group.objectIds.includes(netId)) ||
-    draft.constraints.some((constraint) =>
-      constraint.objectIds.includes(netId),
-    ) ||
-    draft.instances.some((instance) => instance.mosBulkBinding?.netId === netId)
-  ) {
-    return;
-  }
-  const referencedEvidence = draft.connectivityEvidence.filter((evidence) =>
-    connectivityEvidenceNetIds(evidence).includes(netId),
-  );
-  if (
-    referencedEvidence.some((evidence) =>
-      options.protectedEvidenceIds?.has(evidence.id),
-    ) ||
-    (options.deferInto && referencedEvidence.length > 0)
-  ) {
-    return;
-  }
-  const retainedEvidence: typeof draft.connectivityEvidence = [];
-  for (const evidence of draft.connectivityEvidence) {
-    if (evidence.kind === "explicit-equivalence") {
-      if (!evidence.memberNetIds.includes(netId)) {
-        retainedEvidence.push(evidence);
-        continue;
-      }
-      const memberNetIds = evidence.memberNetIds.filter(
-        (memberNetId) => memberNetId !== netId,
-      );
-      changedObjectIds.add(evidence.id);
-      if (memberNetIds.length >= 2) {
-        evidence.memberNetIds = memberNetIds;
-        retainedEvidence.push(evidence);
-      }
-      continue;
-    }
-    if (evidence.netId === netId) {
-      changedObjectIds.add(evidence.id);
-      continue;
-    }
-    retainedEvidence.push(evidence);
-  }
-  draft.connectivityEvidence = retainedEvidence;
-  let clearedBulkDefault = false;
-  if (draft.mosBulkDefaults?.nmosNetId === netId) {
-    delete draft.mosBulkDefaults.nmosNetId;
-    clearedBulkDefault = true;
-  }
-  if (draft.mosBulkDefaults?.pmosNetId === netId) {
-    delete draft.mosBulkDefaults.pmosNetId;
-    clearedBulkDefault = true;
-  }
-  if (clearedBulkDefault) {
-    if (
-      !draft.mosBulkDefaults?.nmosNetId &&
-      !draft.mosBulkDefaults?.pmosNetId
-    ) {
-      delete draft.mosBulkDefaults;
-    }
-    changedObjectIds.add(draft.id);
-  }
-  draft.nets = draft.nets.filter((candidate) => candidate.id !== netId);
-  changedObjectIds.add(netId);
-}
-
-/**
- * A Cell bulk default may deliberately point at an ordinary custom body-bias
- * Net, so a default is not invalid merely because it is not a power Net.
- * However, when the default was a reviewed supply before this transaction and
- * the transaction removes its final supply claim, retaining the materialized
- * B terminals turns deleted presentation into stale electrical truth.
- *
- * Compare the transaction boundary rather than guessing from the final shape:
- * explicit custom defaults remain untouched, while a last VDD/Ground marker
- * deletion revokes only the bindings that were materialized from that default.
- */
-function revokeInvalidatedSupplyBulkDefaults(
-  before: SchematicDocument,
-  draft: SchematicDocument,
-  changedObjectIds: Set<string>,
-  deferNetPrune: (netId: string) => void,
-): boolean {
-  const beforeLogical = resolveDocumentLogicalNets(before);
-  const afterLogical = resolveDocumentLogicalNets(draft);
-  let changed = false;
-
-  for (const [kind, field, expectedDomain] of [
-    ["nmos", "nmosNetId", "ground"],
-    ["pmos", "pmosNetId", "vdd"],
-  ] as const) {
-    const beforeDefaultId = before.mosBulkDefaults?.[field];
-    const afterDefaultId = draft.mosBulkDefaults?.[field];
-    if (!beforeDefaultId || !afterDefaultId) continue;
-    const beforeGroup = beforeLogical.byBaseNetId.get(beforeDefaultId);
-    const wasSupply = beforeGroup?.powerDomain === expectedDomain;
-    const remainsSupply =
-      afterLogical.byBaseNetId.get(afterDefaultId)?.powerDomain ===
-      expectedDomain;
-    if (!wasSupply || remainsSupply) continue;
-
-    const replacementGroups = afterLogical.groups.filter(
-      (group) =>
-        group.powerDomain === expectedDomain &&
-        beforeGroup?.name &&
-        group.name &&
-        foldNetName(group.name) === foldNetName(beforeGroup.name),
-    );
-    const replacementNetId =
-      replacementGroups.length === 1 ? replacementGroups[0]!.id : undefined;
-
-    const affectedNetIds = new Set<string>();
-    for (const instance of draft.instances) {
-      if (
-        instance.symbolId !== kind ||
-        instance.mosBulkBinding?.origin !== "cell-default" ||
-        instance.mosBulkBinding.netId !== afterDefaultId
-      ) {
-        continue;
-      }
-      const net = draft.nets.find(
-        (candidate) => candidate.id === afterDefaultId,
-      );
-      if (net && replacementNetId !== afterDefaultId) {
-        net.terminals = net.terminals.filter(
-          (terminal) =>
-            terminal.instanceId !== instance.id || terminal.pinName !== "B",
-        );
-        affectedNetIds.add(net.id);
-        changedObjectIds.add(net.id);
-      }
-      if (replacementNetId) {
-        const replacement = draft.nets.find(
-          (candidate) => candidate.id === replacementNetId,
-        );
-        if (
-          replacement &&
-          !replacement.terminals.some(
-            (terminal) =>
-              terminal.instanceId === instance.id && terminal.pinName === "B",
-          )
-        ) {
-          replacement.terminals.push({ instanceId: instance.id, pinName: "B" });
-          changedObjectIds.add(replacement.id);
-        }
-        instance.mosBulkBinding.netId = replacementNetId;
-      } else {
-        delete instance.mosBulkBinding;
-      }
-      changedObjectIds.add(instance.id);
-    }
-
-    if (replacementNetId) draft.mosBulkDefaults![field] = replacementNetId;
-    else delete draft.mosBulkDefaults![field];
-    changedObjectIds.add(draft.id);
-    changed = true;
-    for (const netId of affectedNetIds) {
-      deferNetPrune(netId);
-    }
-  }
-
-  if (
-    draft.mosBulkDefaults &&
-    !draft.mosBulkDefaults.nmosNetId &&
-    !draft.mosBulkDefaults.pmosNetId
-  ) {
-    delete draft.mosBulkDefaults;
-  }
-  return changed;
-}
-
-function implicitBulkPresentation(
-  instance: SchematicDocument["instances"][number],
-  resolver: SymbolResolver | undefined,
-): boolean {
-  const resolved = resolver?.resolve(
-    instance.symbolId,
-    instance.symbolVariantId,
-  );
-  return Boolean(
-    resolved?.variant?.hiddenPinNames.includes("B") ||
-    resolved?.definition.pins.find((pin) => pin.name === "B")?.presentation
-      .visibility === "implicit",
-  );
-}
-
-/**
- * A materialized cell-default body is policy-owned, not route-owned.  Route
- * splitting may temporarily place its terminal on a detached Base Net, but it
- * must converge back to the currently configured default before validation.
- * An explicit disconnect first removes mosBulkBinding, so explicit four-pin
- * body editing remains outside this invariant.
- */
-function reconcileMaterializedMosBulkBindings(
-  draft: SchematicDocument,
-  changedObjectIds: Set<string>,
-  deferNetPrune: (netId: string) => void,
-): boolean {
-  let changed = false;
-  for (const instance of draft.instances) {
-    if (instance.mosBulkBinding?.origin !== "cell-default") continue;
-    if (hasExplicitMosBulkRoute(draft, instance.id)) {
-      delete instance.mosBulkBinding;
-      changedObjectIds.add(instance.id);
-      changed = true;
-      continue;
-    }
-    const kind = mosBulkKind(instance);
-    const targetNetId =
-      kind === "nmos"
-        ? draft.mosBulkDefaults?.nmosNetId
-        : kind === "pmos"
-          ? draft.mosBulkDefaults?.pmosNetId
-          : undefined;
-    const currentNets = draft.nets.filter((net) =>
-      net.terminals.some(
-        (terminal) =>
-          terminal.instanceId === instance.id && terminal.pinName === "B",
-      ),
-    );
-    const target = targetNetId
-      ? draft.nets.find((net) => net.id === targetNetId)
-      : undefined;
-
-    if (!target) {
-      for (const net of currentNets) {
-        net.terminals = net.terminals.filter(
-          (terminal) =>
-            terminal.instanceId !== instance.id || terminal.pinName !== "B",
-        );
-        changedObjectIds.add(net.id);
-        deferNetPrune(net.id);
-      }
-      delete instance.mosBulkBinding;
-      changedObjectIds.add(instance.id);
-      changed = true;
-      continue;
-    }
-
-    for (const net of currentNets) {
-      if (net.id === target.id) continue;
-      net.terminals = net.terminals.filter(
-        (terminal) =>
-          terminal.instanceId !== instance.id || terminal.pinName !== "B",
-      );
-      changedObjectIds.add(net.id);
-      deferNetPrune(net.id);
-      changed = true;
-    }
-    if (
-      !target.terminals.some(
-        (terminal) =>
-          terminal.instanceId === instance.id && terminal.pinName === "B",
-      )
-    ) {
-      target.terminals.push({ instanceId: instance.id, pinName: "B" });
-      changedObjectIds.add(target.id);
-      changed = true;
-    }
-    if (instance.mosBulkBinding.netId !== target.id) {
-      instance.mosBulkBinding.netId = target.id;
-      changedObjectIds.add(instance.id);
-      changed = true;
-    }
-  }
-  return changed;
-}
-
-interface BulkDefaultIdentity {
-  name?: string;
-  scope?: "local" | "global";
-  powerDomain?: "ground" | "vdd";
-}
-
-function retargetMosBulkDefaultsAfterSplit(
-  draft: SchematicDocument,
-  originalNetId: string,
-  splitNetIds: readonly string[],
-  identity: BulkDefaultIdentity | undefined,
-  changedObjectIds: Set<string>,
-): boolean {
-  if (!identity || (!identity.name && !identity.powerDomain)) return false;
-  const resolution = resolveDocumentLogicalNets(draft);
-  const matchingGroups = [
-    ...new Map(
-      splitNetIds.flatMap((netId) => {
-        const group = resolution.byBaseNetId.get(netId);
-        if (!group) return [];
-        if (
-          identity.name &&
-          (!group.name ||
-            foldNetName(group.name) !== foldNetName(identity.name))
-        ) {
-          return [];
-        }
-        if (identity.scope && group.scope !== identity.scope) return [];
-        if (
-          identity.powerDomain &&
-          group.powerDomain !== identity.powerDomain
-        ) {
-          return [];
-        }
-        return [[group.id, group] as const];
-      }),
-    ).values(),
-  ];
-  if (matchingGroups.length !== 1) return false;
-
-  const matchingBaseNetIds = matchingGroups[0]!.baseNetIds
-    .filter((netId) => splitNetIds.includes(netId))
-    .sort((left, right) => left.localeCompare(right, "en"));
-  const targetNetId = matchingBaseNetIds[0];
-  if (!targetNetId) return false;
-
-  let changed = false;
-  if (
-    draft.mosBulkDefaults?.nmosNetId === originalNetId &&
-    targetNetId !== originalNetId
-  ) {
-    draft.mosBulkDefaults.nmosNetId = targetNetId;
-    changed = true;
-  }
-  if (
-    draft.mosBulkDefaults?.pmosNetId === originalNetId &&
-    targetNetId !== originalNetId
-  ) {
-    draft.mosBulkDefaults.pmosNetId = targetNetId;
-    changed = true;
-  }
-  if (changed) changedObjectIds.add(draft.id);
-  return changed;
 }
 
 export function executeTransaction(
@@ -2199,153 +1343,19 @@ export function executeTransaction(
         }
         break;
       }
-      case "add_cell_terminal": {
-        if (!draft.netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Document has no formal Cell interface",
-          );
-        }
-        const terminal = CellNetlistTerminalSchema.parse(edit.terminal);
-        if (
-          draft.netlist.terminals.some(
-            (candidate) => candidate.id === terminal.id,
-          )
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Cell terminal already exists: ${terminal.id}`,
-          );
-        }
-        const index = edit.index ?? draft.netlist.terminals.length;
-        if (index > draft.netlist.terminals.length) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Cell terminal index ${index} exceeds interface length ${draft.netlist.terminals.length}`,
-          );
-        }
-        draft.netlist.terminals.splice(index, 0, terminal);
-        changedObjectIds.add(terminal.id);
-        connectivityChanged = true;
-        break;
-      }
-      case "update_cell_terminal": {
-        const terminal = draft.netlist?.terminals.find(
-          (candidate) => candidate.id === edit.terminalId,
-        );
-        if (!terminal) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Cell terminal does not exist: ${edit.terminalId}`,
-          );
-        }
-        if (edit.name !== undefined) {
-          terminal.name = edit.name;
-          for (const annotation of draft.annotations) {
-            if (
-              annotation.binding?.kind === "cell-terminal-name" &&
-              annotation.binding.terminalId === terminal.id &&
-              annotation.formatOverride
-            ) {
-              delete annotation.formatOverride;
-              changedObjectIds.add(annotation.id);
-            }
-          }
-        }
-        if (edit.direction !== undefined) terminal.direction = edit.direction;
-        if (edit.interfaceInstanceIds !== undefined) {
-          terminal.interfaceInstanceIds = [...edit.interfaceInstanceIds];
-        }
-        changedObjectIds.add(terminal.id);
-        connectivityChanged = true;
-        break;
-      }
-      case "remove_cell_terminal": {
-        const index =
-          draft.netlist?.terminals.findIndex(
-            (candidate) => candidate.id === edit.terminalId,
-          ) ?? -1;
-        if (index < 0 || !draft.netlist) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Cell terminal does not exist: ${edit.terminalId}`,
-          );
-        }
-        const [removedTerminal] = draft.netlist.terminals.splice(index, 1);
-        if (draft.presentation.cellSymbol?.pinPlacements) {
-          const retained = draft.presentation.cellSymbol.pinPlacements.filter(
-            (placement) => placement.terminalId !== edit.terminalId,
-          );
-          if (
-            retained.length !==
-            draft.presentation.cellSymbol.pinPlacements.length
-          ) {
-            draft.presentation.cellSymbol = {
-              ...draft.presentation.cellSymbol,
-              ...(retained.length > 0 ? { pinPlacements: retained } : {}),
-            };
-            if (retained.length === 0) {
-              delete draft.presentation.cellSymbol.pinPlacements;
-            }
-          }
-        }
-        changedObjectIds.add(edit.terminalId);
-        if (removedTerminal) {
-          deferNetPrune(removedTerminal.netId);
-        }
-        connectivityChanged = true;
-        break;
-      }
-      case "reorder_cell_terminals": {
-        if (!draft.netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Document has no formal Cell interface",
-          );
-        }
-        const currentIds = draft.netlist.terminals.map(
-          (terminal) => terminal.id,
-        );
-        if (
-          edit.terminalIds.length !== currentIds.length ||
-          new Set(edit.terminalIds).size !== currentIds.length ||
-          currentIds.some((id) => !edit.terminalIds.includes(id))
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Cell terminal order must contain every existing terminal exactly once",
-          );
-        }
-        const terminalById = new Map(
-          draft.netlist.terminals.map((terminal) => [terminal.id, terminal]),
-        );
-        draft.netlist.terminals = edit.terminalIds.map((id) =>
-          terminalById.get(id)!,
-        );
-        for (const id of edit.terminalIds) changedObjectIds.add(id);
-        connectivityChanged = true;
-        break;
-      }
+      case "add_cell_terminal":
+      case "update_cell_terminal":
+      case "remove_cell_terminal":
+      case "reorder_cell_terminals":
       case "set_cell_formal_parameters": {
-        if (!draft.netlist) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Document has no formal Cell interface",
-          );
-        }
-        const seen = new Set<string>();
-        for (const parameter of edit.formalParameters) {
-          const folded = parameter.name.toLowerCase();
-          if (seen.has(folded)) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Cell formal parameter ${parameter.name} is duplicated under case folding`,
-            );
-          }
-          seen.add(folded);
-        }
-        draft.netlist.formalParameters = structuredClone(edit.formalParameters);
-        changedObjectIds.add(draft.id);
+        const outcome = applyCellInterfaceEdit(edit, {
+          draft,
+          changedObjectIds,
+          deferNetPrune,
+          reject: rejectAt,
+        });
+        if (!outcome.ok) return outcome.rejection;
+        if (outcome.connectivityChanged) connectivityChanged = true;
         break;
       }
       case "set_route_points": {
@@ -3566,289 +2576,26 @@ export function executeTransaction(
         connectivityChanged = true;
         break;
       }
-      case "set_presentation_style": {
-        draft.presentation.styleProfileId = edit.styleProfileId;
-        if (edit.styleOverrides === null) {
-          delete draft.presentation.styleOverrides;
-        } else if (edit.styleOverrides !== undefined) {
-          draft.presentation.styleOverrides = structuredClone(
-            edit.styleOverrides,
-          );
-        }
-        changedObjectIds.add(draft.id);
-        break;
-      }
-      case "set_cell_symbol_presentation": {
-        if (edit.presentation === null) {
-          delete draft.presentation.cellSymbol;
-        } else {
-          draft.presentation.cellSymbol = structuredClone(edit.presentation);
-        }
-        changedObjectIds.add(draft.id);
-        break;
-      }
-      case "upsert_schematic_annotation": {
-        const existingIndex = draft.annotations.findIndex(
-          (annotation) => annotation.id === edit.annotation.id,
-        );
-        const existing = draft.annotations[existingIndex];
-        if (existing?.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Annotation is locked: ${existing.id}`,
-          );
-        }
-        const annotation = AnnotationSchema.parse(edit.annotation);
-        const bindingError = validateNetLabelBinding(draft, annotation);
-        if (bindingError) return rejectAt("EDIT_PRECONDITION", bindingError);
-        if (existingIndex >= 0) draft.annotations[existingIndex] = annotation;
-        else draft.annotations.push(annotation);
-        changedObjectIds.add(annotation.id);
-        break;
-      }
-      case "remove_schematic_annotation": {
-        const index = draft.annotations.findIndex(
-          (annotation) => annotation.id === edit.annotationId,
-        );
-        const annotation = draft.annotations[index];
-        if (!annotation) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Annotation does not exist: ${edit.annotationId}`,
-          );
-        }
-        if (annotation.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Annotation is locked: ${annotation.id}`,
-          );
-        }
-        if (
-          [...draft.layoutGroups, ...draft.constraints].some((item) =>
-            item.objectIds.includes(annotation.id),
-          )
-        ) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Annotation is referenced by layout intent: ${annotation.id}`,
-          );
-        }
-        const ownerNetIds = removeConnectivityEvidenceOwnedBy(
-          draft,
-          new Set([annotation.id]),
-          changedObjectIds,
-        );
-        draft.annotations.splice(index, 1);
-        changedObjectIds.add(annotation.id);
-        for (const netId of ownerNetIds) {
-          deferNetPrune(netId);
-        }
-        if (ownerNetIds.length > 0) connectivityChanged = true;
-        break;
-      }
-      case "upsert_drafting_object": {
-        ensureDraftingLayer(draft);
-        const objects = draft.drafting!.objects;
-        const existingIndex = objects.findIndex(
-          (item) => item.id === edit.object.id,
-        );
-        const existing = objects[existingIndex];
-        const parsed = DraftingObjectSchema.parse(edit.object);
-        // Locking protects the object from normal replacement, but it must not
-        // become irreversible. Permit exactly one exceptional replacement: the
-        // same canonical object with only `locked` switched to false. Parsing
-        // both sides through the model schema gives this comparison a stable
-        // persisted shape and prevents Unlock from smuggling in another edit.
-        const isPureUnlock =
-          existing?.locked === true &&
-          parsed.locked === false &&
-          JSON.stringify({ ...existing, locked: false }) ===
-            JSON.stringify(parsed);
-        if (existing?.locked && !isPureUnlock) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Drafting object is locked: ${existing.id}`,
-          );
-        }
-        if (parsed.kind === "floating-symbol") {
-          if (!resolver) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `A Symbol Resolver is required to validate a floating symbol: ${parsed.symbolId}`,
-            );
-          }
-          const resolved = resolver.resolve(parsed.symbolId);
-          if (!resolved) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Unknown floating symbol: ${parsed.symbolId}`,
-            );
-          }
-          // ADR 0010: a floating symbol must reference a `decorative: true`
-          // catalog entry whose definition contains no terminal, enforced here
-          // via the Symbol Resolver (the model schema cannot check the catalog).
-          if (!resolved.definition.decorative) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Floating symbol must be decorative: ${parsed.symbolId}`,
-            );
-          }
-          if (resolved.definition.pins.length > 0) {
-            return rejectAt(
-              "EDIT_PRECONDITION",
-              `Floating symbol must be terminal-free: ${parsed.symbolId}`,
-            );
-          }
-        }
-        if (existingIndex >= 0) objects[existingIndex] = parsed;
-        else objects.push(parsed);
-        changedObjectIds.add(parsed.id);
-        break;
-      }
-      case "remove_drafting_object": {
-        ensureDraftingLayer(draft);
-        const objects = draft.drafting!.objects;
-        const index = objects.findIndex((item) => item.id === edit.objectId);
-        const object = objects[index];
-        if (!object) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Drafting object does not exist: ${edit.objectId}`,
-          );
-        }
-        objects.splice(index, 1);
-        changedObjectIds.add(object.id);
-        break;
-      }
-      case "set_layout_group": {
-        const index = draft.layoutGroups.findIndex(
-          (group) => group.id === edit.group.id,
-        );
-        const existing = draft.layoutGroups[index];
-        if (existing?.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Layout group is locked: ${existing.id}`,
-          );
-        }
-        const group = LayoutGroupSchema.parse(edit.group);
-        if (index >= 0) draft.layoutGroups[index] = group;
-        else draft.layoutGroups.push(group);
-        changedObjectIds.add(group.id);
-        break;
-      }
-      case "remove_layout_group": {
-        const index = draft.layoutGroups.findIndex(
-          (group) => group.id === edit.groupId,
-        );
-        const group = draft.layoutGroups[index];
-        if (!group) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Layout group does not exist: ${edit.groupId}`,
-          );
-        }
-        if (group.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Layout group is locked: ${group.id}`,
-          );
-        }
-        draft.layoutGroups.splice(index, 1);
-        changedObjectIds.add(group.id);
-        break;
-      }
-      case "set_layout_constraint": {
-        const index = draft.constraints.findIndex(
-          (constraint) => constraint.id === edit.constraint.id,
-        );
-        const existing = draft.constraints[index];
-        if (existing?.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Layout constraint is locked: ${existing.id}`,
-          );
-        }
-        const constraint = LayoutConstraintSchema.parse(edit.constraint);
-        if (index >= 0) draft.constraints[index] = constraint;
-        else draft.constraints.push(constraint);
-        changedObjectIds.add(constraint.id);
-        break;
-      }
-      case "remove_layout_constraint": {
-        const index = draft.constraints.findIndex(
-          (constraint) => constraint.id === edit.constraintId,
-        );
-        const constraint = draft.constraints[index];
-        if (!constraint) {
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Layout constraint does not exist: ${edit.constraintId}`,
-          );
-        }
-        if (constraint.locked) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Layout constraint is locked: ${constraint.id}`,
-          );
-        }
-        draft.constraints.splice(index, 1);
-        changedObjectIds.add(constraint.id);
-        break;
-      }
+      case "set_presentation_style":
+      case "set_cell_symbol_presentation":
+      case "upsert_schematic_annotation":
+      case "remove_schematic_annotation":
+      case "upsert_drafting_object":
+      case "remove_drafting_object":
+      case "set_layout_group":
+      case "remove_layout_group":
+      case "set_layout_constraint":
+      case "remove_layout_constraint":
       case "align_instances": {
-        if (new Set(edit.instanceIds).size !== edit.instanceIds.length) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Alignment instance IDs must be unique",
-          );
-        }
-        const instances = edit.instanceIds.map((id) =>
-          draft.instances.find((instance) => instance.id === id),
-        );
-        if (instances.some((instance) => !instance)) {
-          const missing = edit.instanceIds.find(
-            (id) => !draft.instances.some((instance) => instance.id === id),
-          );
-          return rejectAt(
-            "OBJECT_NOT_FOUND",
-            `Instance does not exist: ${missing}`,
-          );
-        }
-        const lockedInstanceId = edit.instanceIds.find((id) =>
-          lockedLayoutOwner(draft, id),
-        );
-        if (lockedInstanceId) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            `Instance ${lockedInstanceId} is locked by layout intent ${lockedLayoutOwner(draft, lockedInstanceId)}`,
-          );
-        }
-        if (instances.some((instance) => instance!.placement === null)) {
-          return rejectAt(
-            "EDIT_PRECONDITION",
-            "Every aligned instance must be placed",
-          );
-        }
-        const coordinate =
-          edit.coordinate ?? instances[0]!.placement!.position[edit.axis];
-        for (const instance of instances) {
-          const oldCoordinate = instance!.placement!.position[edit.axis];
-          instance!.placement!.position[edit.axis] = coordinate;
-          for (const annotation of draft.annotations) {
-            if (
-              annotation.anchor.kind === "object" &&
-              annotation.anchor.objectId === instance!.id
-            ) {
-              translateObjectAnchoredAnnotation(annotation, instance!.id, {
-                x: edit.axis === "x" ? coordinate - oldCoordinate : 0,
-                y: edit.axis === "y" ? coordinate - oldCoordinate : 0,
-              });
-              changedObjectIds.add(annotation.id);
-            }
-          }
-          changedObjectIds.add(instance!.id);
-        }
+        const outcome = applyPresentationLayoutEdit(edit, {
+          draft,
+          resolver,
+          changedObjectIds,
+          deferNetPrune,
+          reject: rejectAt,
+        });
+        if (!outcome.ok) return outcome.rejection;
+        if (outcome.connectivityChanged) connectivityChanged = true;
         break;
       }
     }
